Added cache manager. Fixes #96.
